Moved ranking to proper path (/ranking)
