+++ music_karaoke/templates/layout/navbar.html
@@ -15,7 +15,7 @@
     <ul class="list">
       {% if user.is_authenticated %}
         <li class="list__item"><a href="{% url 'songs:index' %}">Songs</a></li>
-        <li class="list__item"><a href="{% url 'songs:ranking' %}">Ranking</a></li>
+        <li class="list__item"><a href="{% url 'ranking' %}">Ranking</a></li>
         <li class="list__item"><a href="">Favourites</a></li>
         <li class="list__item"><a href="logout/">Logout</a></li>
       {% else %}
